split storage integration tests
